backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.20.*
numpy
pandas
scikit-learn
joblib
matplotlib
pydantic
flask